+++ packages/eslint-config/library.js
@@ -40,10 +40,15 @@ module.exports = {
   rules: {
     // Allow unused vars/args that start with _
     "@typescript-eslint/no-unused-vars": [
-      "warn",
+      "error",
       {
+        args: "all",
         argsIgnorePattern: "^_",
+        caughtErrors: "all",
+        caughtErrorsIgnorePattern: "^_",
+        destructuredArrayIgnorePattern: "^_",
         varsIgnorePattern: "^_",
+        ignoreRestSiblings: true,
       },
     ],
   },