+++ packages/react-core/src/index.ts
@@ -1,5 +1,7 @@
 export * from "./modules/core";
 export * from "./modules/feed";
 export * from "./modules/in-app-messages";
+export * from "./modules/ms-teams";
 export * from "./modules/slack";
 export * from "./modules/i18n";
+export * from "./interfaces";