+++ packages/react-core/src/interfaces.ts
@@ -1,0 +1,4 @@
+export type RecipientObject = {
+  objectId: string;
+  collection: string;
+};