+++ packages/react/src/index.ts
@@ -3,6 +3,7 @@ import "./theme.css";
 export * from "./modules/core";
 export * from "./modules/feed";
 export * from "./modules/in-app-messages";
+export * from "./modules/ms-teams";
 export * from "./modules/slack";
 
 export * from "@knocklabs/react-core";