+++ packages/react/src/modules/ms-teams/index.ts
@@ -1,0 +1,3 @@
+export * from "./components/MsTeamsAuthButton";
+export * from "./components/MsTeamsAuthContainer";
+export * from "./components/MsTeamsChannelCombobox";