+++ packages/react/src/modules/slack/components/SlackAuthButton/SlackAuthButton.tsx
@@ -7,6 +7,7 @@ import {
 import { FunctionComponent } from "react";
 import { useEffect } from "react";
 
+import { openPopupWindow } from "../../../core/utils";
 import "../../theme.css";
 import { SlackIcon } from "../SlackIcon";
 
@@ -18,29 +19,6 @@ export interface SlackAuthButtonProps {
   onAuthenticationComplete?: (authenticationResp: string) => void;
   additionalScopes?: string[];
 }
-
-const openSlackOauthPopup = (url: string) => {
-  const width = 600;
-  const height = 800;
-  // Calculate the position to center the window
-  const screenLeft = window.screenLeft ?? window.screenX;
-  const screenTop = window.screenTop ?? window.screenY;
-
-  const innerWidth =
-    window.innerWidth ?? document.documentElement.clientWidth ?? screen.width;
-  const innerHeight =
-    window.innerHeight ??
-    document.documentElement.clientHeight ??
-    screen.height;
-
-  const left = innerWidth / 2 - width / 2 + screenLeft;
-  const top = innerHeight / 2 - height / 2 + screenTop;
-
-  // Window features
-  const features = `width=${width},height=${height},top=${top},left=${left}`;
-
-  window.open(url, "Slack OAuth", features);
-};
 
 export const SlackAuthButton: FunctionComponent<SlackAuthButtonProps> = ({
   slackClientId,
@@ -83,7 +61,7 @@ export const SlackAuthButton: FunctionComponent<SlackAuthButtonProps> = ({
         if (onAuthenticationComplete) {
           onAuthenticationComplete(event.data);
         }
-      } catch (error) {
+      } catch (_error) {
         setConnectionStatus("error");
       }
     };
@@ -120,7 +98,7 @@ export const SlackAuthButton: FunctionComponent<SlackAuthButtonProps> = ({
   if (connectionStatus === "error") {
     return (
       <button
-        onClick={() => openSlackOauthPopup(buildSlackAuthUrl())}
+        onClick={() => openPopupWindow(buildSlackAuthUrl())}
         className="rsk-connect__button rsk-connect__button--error"
         onMouseEnter={() => setActionLabel(reconnectLabel)}
         onMouseLeave={() => setActionLabel(null)}
@@ -137,7 +115,7 @@ export const SlackAuthButton: FunctionComponent<SlackAuthButtonProps> = ({
   if (connectionStatus === "disconnected") {
     return (
       <button
-        onClick={() => openSlackOauthPopup(buildSlackAuthUrl())}
+        onClick={() => openPopupWindow(buildSlackAuthUrl())}
         className="rsk-connect__button rsk-connect__button--disconnected"
       >
         <SlackIcon height="16px" width="16px" />