+++ examples/expo-example/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @knocklabs/expo-example
 
+## 0.1.7
+
+### Patch Changes
+
+- @knocklabs/expo@0.2.7
+
+## 0.1.6
+
+### Patch Changes
+
+- @knocklabs/expo@0.2.6
+
+## 0.1.5
+
+### Patch Changes
+
+- @knocklabs/expo@0.2.5
+
+## 0.1.4
+
+### Patch Changes
+
+- @knocklabs/expo@0.2.4
+
+## 0.1.3
+
+### Patch Changes
+
+- @knocklabs/expo@0.2.3
+
+## 0.1.2
+
+### Patch Changes
+
+- @knocklabs/expo@0.2.2
+
 ## 0.1.1
 
 ### Patch Changes