+++ packages/expo/CHANGELOG.md
@@ -1,5 +1,67 @@
 # @knocklabs/expo
 
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies [85418a0]
+- Updated dependencies [2161d3f]
+- Updated dependencies [8cc9338]
+- Updated dependencies [da84a75]
+- Updated dependencies [2161d3f]
+- Updated dependencies [2161d3f]
+- Updated dependencies [1ba1393]
+- Updated dependencies [b4b5c02]
+  - @knocklabs/react-core@0.3.2
+  - @knocklabs/client@0.11.2
+  - @knocklabs/react-native@0.5.7
+
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [b9f6712]
+  - @knocklabs/react-core@0.3.1
+  - @knocklabs/client@0.11.1
+  - @knocklabs/react-native@0.5.6
+
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [013ad8d]
+  - @knocklabs/react-core@0.3.0
+  - @knocklabs/client@0.11.0
+  - @knocklabs/react-native@0.5.5
+
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [26db496]
+- Updated dependencies [988aaf9]
+  - @knocklabs/client@0.10.17
+  - @knocklabs/react-core@0.2.29
+  - @knocklabs/react-native@0.5.4
+
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [bc99374]
+  - @knocklabs/client@0.10.16
+  - @knocklabs/react-core@0.2.28
+  - @knocklabs/react-native@0.5.3
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [26166e3]
+  - @knocklabs/client@0.10.15
+  - @knocklabs/react-core@0.2.27
+  - @knocklabs/react-native@0.5.2
+
 ## 0.2.1
 
 ### Patch Changes